+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Alpha 0.3.1] 2025-02-05
+
+### Changed
+- /calendar will now force download the file.
+- /empquery now returns state
+
+---
 ## [Alpha 0.3.0] 2025-02-02
 
 ### Added